Some modifications in XmlVIsitor

+++ tinylang/src/visitor/xmlvisitor/visitbool.cc
@@ -4,7 +4,7 @@ void XmlVisitor::visit(AstBoolLiteralNode *node)
 {
     string tabs = "";
 
-    for (int t = 0; t < d_indent; ++t) 
+    for (size_t t = 0; t < d_indent; ++t) 
         tabs.append("\t");
 
     d_file << boolalpha << tabs << "<Bool>" << node->value() << "</Bool>"
