+++ tinylang/src/visitor/xmlvisitor/visitfor.cc
@@ -4,7 +4,7 @@ void XmlVisitor::visit(AstForStmtNode *node)
 {
     string tabs = "";
 
-    for (int t = 0; t < d_indent; ++t) 
+    for (size_t t = 0; t < d_indent; ++t) 
         tabs.append("\t");
 
     d_file << tabs << "<ForLoop>" << endl;
